MainActivity moved to common package
