+++ app/src/main/res/layout/activity_main.xml
@@ -4,5 +4,5 @@
     android:id="@+id/mainActivityContainer"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.MainActivity">
+    tools:context=".ui.common.MainActivity">
 </FrameLayout>
